backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai>=0.8.0
google-api-python-client
beautifulsoup4==4.12.2
python-dotenv==1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
